cs50w/violin_sheet_music/requirements.txt
Django==5.0.6
Pillow==10.3.0
python-magic-bin==0.4.14